fs: don't use GENERIC_READ with CreateFile when obtaining file info
We don't really need this flag, because we don't do any actual reading
except reading metadata. But using GENERIC_READ on file opened in exclusive
mode, results in 'OSError: [WinError 6] The handle is invalid' being raised,
which is caused by the GetFileInformationByHandle returing 0. This bug
is pretty easy to reproduce by just opening a file in excel and then trying
to feed it to getfileinfo().

This patch drops GENERIC_READ and also provides a test.

Signed-off-by: Ruslan Kuprieiev <kupruser@gmail.com>